requirements.txt
pandas
sqlalchemy  == 2.0.31
python-dotenv
psycopg2-binary
yfinance